working sync
problem with changing variants

+++ server/webhooks/products_Delete.js
@@ -8,27 +8,31 @@ import chalk from 'chalk'
 
 const deleteProductHookHandler = async (topic, shop, webhookRequestBody, webhookId, apiVersion) => {
    try {
-      console.log(chalk.bgCyanBright(topic));
+      console.log(chalk.bgCyanBright(topic))
       const offlineSession = await loadSessionFromStore(shop)
       const { id } = JSON.parse(webhookRequestBody)
       const multiStorePd = await getMultiStoreFromProductId(id)
-      const fullData = []
-      for (const shopiData of multiStorePd.shopifyData) {
-         const session = await getSessionsFromStores([shopiData.store], offlineSession.id, {
-            offline: true,
-            noCurrent: true
-         })
-         if (session && session.length) {
-            fullData.push({ ...shopiData.toJSON(), sessionId: session[0].session.id })
+      if (multiStorePd) {
+         const fullData = []
+         for (const shopiData of multiStorePd.shopifyData) {
+            const session = await getSessionsFromStores([shopiData.store], offlineSession.id, {
+               offline: true,
+               noCurrent: true
+            })
+            if (session && session.length) {
+               fullData.push({ ...shopiData.toJSON(), sessionId: session[0].session.id })
+            }
          }
-      }
 
-      for (const sessionObj of fullData) {
-         const loadedSession = await sessionHandler.loadSession(sessionObj.sessionId)
-         console.log(chalk.blue(`updating product in ${sessionObj.store.shop}...`))
-         await deleteProduct(sessionObj.id, loadedSession)
+         for (const sessionObj of fullData) {
+            const loadedSession = await sessionHandler.loadSession(sessionObj.sessionId)
+            console.log(chalk.blue(`deleting product in ${sessionObj.store.shop}...`))
+            await deleteProduct(sessionObj.id, loadedSession)
+         }
+         await MultiStoreProductModel.findOneAndDelete({ _id: multiStorePd._id })
+      } else {
+         console.log(chalk.yellow('No record of this product being associated to others'))
       }
-      await MultiStoreProductModel.findOneAndDelete({ _id: multiStorePd._id })
       await resetSession(offlineSession.id)
       console.log(chalk.green('done.'))
    } catch (error) {
